requirements.txt
PySimpleGUI
pillow
requests
python-dotenv